+++ 02-first-steps/package.json
@@ -24,6 +24,7 @@
     "@typescript-eslint/eslint-plugin": "7.13.1",
     "@typescript-eslint/parser": "7.13.1",
     "@vitejs/plugin-react-swc": "3.5.0",
+    "@vitest/coverage-v8": "^3.2.4",
     "eslint": "8.57.0",
     "eslint-plugin-react-hooks": "4.6.2",
     "eslint-plugin-react-refresh": "0.4.7",
